add kubernetes deployment
this will run the ml-service in kubernetes, and then run lammps alongside
it from a singularity container. We randomly select parameters x,y,z
from some range and then use that to train each of three models in the server.
We then have another script that does the same to generate testing data,
save the predictions alongside the actual values, and calculate an accuracy.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>